requirements.txt
# Core MuJoCo dependencies
mujoco>=3.0.0
numpy>=1.24.0
matplotlib>=3.7.0

# Deep Learning and RL
torch>=2.0.0
torchvision>=0.15.0
gym>=0.26.0
gymnasium>=0.28.0
stable-baselines3>=2.0.0

# Visualization and rendering
opencv-python>=4.8.0
imageio>=2.31.0
pillow>=10.0.0
PyOpenGL>=3.1.6
glfw>=2.6.0
seaborn>=0.12.0

# Data processing and utilities
pyyaml>=6.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# Logging and monitoring
tensorboard>=2.13.0
wandb>=0.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Development tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
